requirements_local.txt
MetaTrader5
pandas
pandas_ta
google-generativeai
fastapi
uvicorn
python-dotenv
supabase
pydantic
feedparser
